refactor benchmark
