--- benchmark/websocket-benchmark.rb
@@ -1,171 +1,0 @@
-#!/usr/bin/env ruby
-
-require 'logger'
-require 'kontena-websocket-client'
-
-Thread.abort_on_exception = true
-
-log_level = ENV['LOG_LEVEL'] || Logger::WARN
-WEBSOCKET_OPTIONS = {
-  connect_timeout: 1.0,
-  open_timeout: 1.0,
-  ping_timeout: 1.0,
-  ping_interval: nil,
-  write_timeout: 5.0,
-}
-
-$logger = Logger.new(STDERR)
-$logger.level = log_level
-$logger.progname = 'websocket-benchmark'
-
-Kontena::Websocket::Logging.initialize_logger(STDERR, log_level)
-
-def with_rate(rate, duration, &block)
-  t0 = Time.now
-  interval = 1.0 / rate
-  count = 0
-  count_miss = 0
-  total_yield = 0.0
-
-  while (t = Time.now) < t0 + duration
-    yield t
-
-    t_yield = Time.now
-
-    count += 1
-    total_yield += (t_yield - t)
-
-    t_next = t0 + count * interval
-
-    if t_next > t_yield
-      sleep t_next - t_yield
-    else
-      count_miss += 1
-    end
-  end
-
-  t_total = t - t0
-
-  return {
-    time: t_total,
-    count: count,
-    rate: count / t_total,
-    util: total_yield / t_total,
-    miss: (count_miss / count),
-  }
-end
-
-def websocket_benchark_sender(client, rate: 1000, duration: 5.0, message_size: 1000)
-  total_size = 0
-
-  padding = 'X'*(message_size - 15)
-
-  stats = with_rate(rate, duration) do |t|
-    client.send('%15.6f %s' % [t.to_f, padding])
-
-    total_size += message_size
-  end
-
-  $logger.info "write close..."
-
-  client.close()
-
-  return stats.merge(
-    bytes: total_size,
-  )
-end
-
-def websocket_benchmark_reader(client)
-  count = 0
-  bytes = 0
-  latency_total = 0.0
-  t_start = Time.now
-
-  client.read do |message|
-    t = Time.now.to_f
-    t_s, padding = message.split(' ', 2)
-    t_f = t_s.to_f
-
-    count += 1
-    bytes += message.length
-    latency_total += (t - t_f)
-  end
-
-  t_end = Time.now
-
-  $logger.info "read done"
-
-  t_total = t_end - t_start
-
-  return {
-    time: t_total,
-    count: count,
-    rate: count / t_total,
-    bytes: bytes,
-    bytes_rate: bytes / t_total,
-    latency_avg: latency_total / count,
-  }
-end
-
-def websocket_benchmark(url, **options)
-  send_thread = nil
-  read_stats = nil
-
-  Kontena::Websocket::Client.connect(url, **WEBSOCKET_OPTIONS) do |client|
-    $logger.info "connect: #{client}"
-
-    send_thread = Thread.new {
-      websocket_benchark_sender(client, **options)
-    }
-
-    read_stats = websocket_benchmark_reader(client)
-  end
-
-  send_stats = send_thread.value
-
-  return send_stats, read_stats
-end
-
-url = ENV['URL'] || 'ws://localhost:8080/echo'
-
-RATES = [1, 10, 100, 1000, 3000, 5000, 10000]
-rates = (ENV['RATES']&.split&.map{|r| Integer(r)} || RATES)
-duration = (ENV['DURATION'] || 5.0).to_f
-message_size = (ENV['MESSAGE_SIZE'] || 1000).to_i
-
-HEADER = '%5s  %6s/s %9s: send @ %9s/s (%6s%% %6s%%) read @ %9s/s (%6s%%) = %12s/s ~%9s'
-FORMAT = '%5.2fs %6d/s %9d: send @ %9.2f/s (%6.2f%% %6.2f%%) read @ %9.2f/s (%6.2f%%) = %12s/s ~%9.6fs'
-
-def si(val)
-  if val > 10**9
-    '%.3fG' % [val / 10**9]
-  elsif val > 10**6
-    '%.3fM' % [val / 10**6]
-  elsif val > 10**3
-    '%.3fK' % [val / 10**3]
-  else
-    '%.3f ' % [val]
-  end
-end
-
-puts HEADER % ['TIME ', 'RATE', 'COUNT', 'MESSAGES', 'UTIL', 'MISS', 'MESSAGES', 'DROP', 'BYTES', 'LATENCY']
-
-for rate in rates
-  options = {
-    rate: rate,
-    duration: duration,
-    message_size: message_size,
-  }
-
-  $logger.info "benchmark: #{url} #{options}"
-
-  send_stats, read_stats = websocket_benchmark(url, **options)
-
-  drop_ratio = 1.0 - read_stats[:count].to_f / send_stats[:count].to_f
-
-  puts FORMAT % [
-    duration, rate, send_stats[:count],
-    send_stats[:rate], send_stats[:util] * 100.0, send_stats[:miss] * 100.0,
-    read_stats[:rate], drop_ratio * 100.0, si(read_stats[:bytes_rate]), read_stats[:latency_avg],
-  ]
-end
